add confirm on click

+++ view/backend/deletePost.php
@@ -26,8 +26,8 @@ if (isset($_SESSION['nickname'])){
                 </p>
             </div>
             <div>
-                <a href="index.php?action=deletePost&amp;id=<?php echo $dataPosts['id']; ?>">
-                supprimer
+                <a class="text-danger" href="index.php?action=deletePost&amp;id=<?php echo $dataPosts['id']; ?>"onclick="return confirm('attention suppression définitive !')">
+                    supprimer ?
                 </a>
             </div>
         </div>
